03-rag/requirements.txt
flask
phidata
sqlalchemy
pgvectorflask_cors
